Agrego test de registrar mascota, y agregar caracteristicas por un admin

+++ src/main/java/domain/organizacion/Documento.java
@@ -1,6 +1,6 @@
 package domain.organizacion;
 
 public class Documento {
-    private int numero;
+    private float numero;
     private String tipo;
 }
